walk cycle along a circular path:
created a track for each strip to get the option of transitions, as in https://www.youtube.com/watch?v=zDOFg2spySg
added a circle, converted that to a curve, moved x by one and rotated by 90° in edit mode such that camel is at tangent
created follow path for the armature in object mode
removed alread set key frames for the curve eval. time
added key frames for the eval. time to cover 5 walking cycles
adjusted circle radius in edit mode to reduce slipping walk (needs more tweeking because the radius is small compared to the camel